feat: 가상 포인트 엔진 안정화(QuickJS 복구 로직), 통합 테스트(Step 10, 11) 수립 및 멀티 플랫폼 통합 릴리즈 체계 구축

+++ core/shared/include/Platform/PlatformCompat.h
@@ -17,7 +17,7 @@
 // C++17 filesystem 지원 확인
 #if __cplusplus >= 201703L && __has_include(<filesystem>)
     #include <filesystem>
-#include <algorithm>
+    #include <algorithm>
     #define HAS_FILESYSTEM 1
 #else
     #define HAS_FILESYSTEM 0
@@ -539,7 +539,7 @@ public:
 #if PULSEONE_WINDOWS
         return ::GetCurrentThreadId();
 #else
-        return static_cast<uint32_t>(pthread_self());
+        return (uint32_t)(uintptr_t)pthread_self();
 #endif
     }
 };
